# docs/swagger.yaml
definitions:
  app.Pager:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  errcode.Error:
    type: object
  model.Article:
    properties:
      content:
        type: string
      cover:
        type: string
      created:
        type: integer
      created_by:
        type: string
      deleted:
        type: integer
      desc:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      status:
        type: integer
      title:
        type: string
      updated:
        type: integer
      updated_by:
        type: string
    type: object
  model.ArticleSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Article'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
    type: object
  model.Tag:
    properties:
      created:
        type: integer
      created_by:
        type: string
      deleted:
        type: integer
      id:
        type: integer
      is_del:
        type: integer
      name:
        type: string
      status:
        type: integer
      updated:
        type: integer
      updated_by:
        type: string
    type: object
  model.TagSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
    type: object
info:
  contact: { }
  description: gin写的博客
  termsOfService: https://github.com/lackone/gin-blog
  title: 博客
  version: "1.0"
paths:
  /api/v1/articles:
    get:
      parameters:
        - description: 文章名称
          in: query
          name: name
          type: string
        - description: 标签ID
          in: query
          name: tag_id
          type: integer
        - description: 状态
          in: query
          name: status
          type: integer
        - description: 页码
          in: query
          name: page
          type: integer
        - description: 每页数量
          in: query
          name: page_size
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.ArticleSwagger'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取多个文章
    post:
      parameters:
        - description: 标签ID
          in: body
          name: tag_id
          required: true
          schema:
            type: string
        - description: 文章标题
          in: body
          name: title
          required: true
          schema:
            type: string
        - description: 文章简述
          in: body
          name: desc
          schema:
            type: string
        - description: 封面图片
          in: body
          name: cover
          required: true
          schema:
            type: string
        - description: 文章内容
          in: body
          name: content
          required: true
          schema:
            type: string
        - description: 创建者
          in: body
          name: created_by
          required: true
          schema:
            type: integer
        - description: 状态
          in: body
          name: status
          schema:
            type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 创建文章
  /api/v1/articles/{id}:
    delete:
      parameters:
        - description: 文章ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除文章
    get:
      parameters:
        - description: 文章ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取单个文章
    put:
      parameters:
        - description: 标签ID
          in: body
          name: tag_id
          schema:
            type: string
        - description: 文章标题
          in: body
          name: title
          schema:
            type: string
        - description: 文章简述
          in: body
          name: desc
          schema:
            type: string
        - description: 封面图片
          in: body
          name: cover
          schema:
            type: string
        - description: 文章内容
          in: body
          name: content
          schema:
            type: string
        - description: 修改者
          in: body
          name: updated_by
          required: true
          schema:
            type: string
      produces:
        - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Article'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 更新文章
  /api/v1/tags:
    get:
      parameters:
        - description: 标签名称
          in: query
          maxLength: 100
          name: name
          type: string
        - default: 1
          description: 状态
          enum:
            - 0
            - 1
          in: query
          name: status
          type: integer
        - description: 页码
          in: query
          name: page
          type: integer
        - description: 每页数量
          in: query
          name: page_size
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.TagSwagger'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取多个标签
    post:
      parameters:
        - description: 标签名称
          in: body
          maxLength: 100
          minLength: 3
          name: name
          required: true
          schema:
            type: string
        - default: 1
          description: 状态
          enum:
            - 0
            - 1
          in: body
          name: status
          schema:
            type: integer
        - description: 创建者
          in: body
          maxLength: 100
          minLength: 3
          name: created_by
          schema:
            type: string
      produces:
        - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 新增标签
  /api/v1/tags/{id}:
    delete:
      parameters:
        - description: 标签ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除标签
    put:
      parameters:
        - description: 标签ID
          in: path
          name: id
          required: true
          type: integer
        - description: 标签名称
          in: body
          maxLength: 100
          minLength: 3
          name: name
          schema:
            type: string
        - default: 1
          description: 状态
          enum:
            - 0
            - 1
          in: body
          name: status
          schema:
            type: integer
        - description: 修改者
          in: body
          maxLength: 100
          minLength: 3
          name: updated_by
          required: true
          schema:
            type: string
      produces:
        - application/json
      responses:
        "200":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/model.Tag'
            type: array
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 更新标签
swagger: "2.0"